Update DEVToken.sol
With the onlyWhenTransferEnabled modifier and enableTransfer() we don't need to extend PausableToken anymore.

The validDestination modifier is superfluous.  If this was meant to prevent the contract owners to mint to the contract owner address then they can just mint to any other addresses they own anyway.

Renamed mintToken() to mintMany().